Move some test utilities to test_utils.go
